+++ deltafq/adapters/data/__init__.py
@@ -1,4 +1,5 @@
+from .baostock_gateway import BaostockDataGateway
 from .miniqmt_gateway import MiniQmtDataGateway
 from .yfinance_gateway import YFinanceDataGateway
 
-__all__ = ["YFinanceDataGateway", "MiniQmtDataGateway"]
+__all__ = ["BaostockDataGateway", "YFinanceDataGateway", "MiniQmtDataGateway"]